requirements.txt
# API Framework
fastapi
uvicorn[standard]
python-multipart

# Database
motor
pymongo
pydantic[email]
pydantic-settings

# Background Tasks
celery[redis]
redis

# ML & NLP
transformers
torch
spacy
yake
wordcloud
Pillow
langdetect

# Cloud/Object Storage
boto3

# Utilities
python-dotenv